+++ app/models/homebrew/speciality.rb
@@ -2,6 +2,8 @@
 
 class Homebrew
   class Speciality < ApplicationRecord
+    include Discard::Model
+
     belongs_to :user, touch: :homebrew_updated_at
   end
 end