+++ app/models/homebrew/subclass.rb
@@ -2,6 +2,8 @@
 
 class Homebrew
   class Subclass < ApplicationRecord
+    include Discard::Model
+
     belongs_to :user, touch: :homebrew_updated_at
 
     has_many :homebrew_book_items, class_name: 'Homebrew::Book::Item', as: :itemable, dependent: :destroy