+++ spec/controllers/homebrews/daggerheart/communities_controller_spec.rb
@@ -45,10 +45,10 @@ describe Homebrews::Daggerheart::CommunitiesController do
             character.save
           end
 
-          it 'returns error', :aggregate_failures do
+          it 'discards homebrew', :aggregate_failures do
             expect { request }.not_to change(Daggerheart::Homebrew::Community, :count)
-            expect(response).to have_http_status :unprocessable_content
-            expect(response.parsed_body['errors']).to eq({ 'base' => ['Персонаж с таким обществом существует'] })
+            expect(response).to have_http_status :ok
+            expect(homebrew.reload.discarded?).to be_truthy
           end
         end
       end